validation/requirements.txt
pyyaml
sqlalchemy
psycopg2
python-dotenv
matplotlib
pandas
numpy